Confirmed the simulation ran at hi-res. Might want to move to double precision for symmetry

+++ dynamics/awfl/const.h
@@ -27,7 +27,7 @@ using yakl::COLON;
   #define NGLL 3
 #endif
 
-typedef double real;
+typedef float real;
 
 YAKL_INLINE real constexpr operator"" _fp( long double x ) {
   return static_cast<real>(x);
